refactor(macos): split terminal host view adapters
